Extract products validation to @voyantjs/products-contracts + complete admin products writes
Extraction (mirrors bookings/finance/crm/legal):
- Move the products validation cluster (validation, -core, -public, -shared,
  -config, -content, -catalog) out of the runtime @voyantjs/products package
  into the pure @voyantjs/products-contracts. Its only external imports — the
  two @voyantjs/db helpers (booleanQueryParam, typeIdSchema) — repoint to
  @voyantjs/schema-kit, so the package stays zero-runtime (zod + schema-kit).
- The moved files become one-line re-export stubs in @voyantjs/products, so
  every existing import path keeps working unchanged
  (@voyantjs/products/validation, /public-validation, internal ./validation-*).
- products-contracts exports the seven validation subpaths; gains a schema-kit
  dependency.

Admin SDK (completes the products surface, #1411):
- admin-contracts: add products.create/update/delete deriving from
  insertProductSchema/updateProductSchema, and products.list now derives from
  productListQuerySchema — all from the newly-pure @voyantjs/products-contracts.
  Removes the "writes deferred" note.
- admin-client: typed products.create/update/delete methods.
- Tests: registry covers the products write ids + scopes; a client test creates
  a product (input derived, route defaults applied).
